converted about_zulip.js to typescript
